Fix: CreateGroupPage 완료 버튼 정렬 문제 수정
