+++ src/App.tsx
@@ -2,7 +2,7 @@ import Layout from "@layout/Layout";
 import ChatListPage from "@pages/ChatListPage/page";
 import ChatListSearchPage from "@pages/ChatListSearchPage/page";
 import ChattingPage from "@pages/ChattingPage/page";
-import CreateGroupPage from "@pages/CreateGroupPage/page";
+import { default as CreateGroupPage, default as GroupPage } from "@pages/CreateGroupPage/page";
 import CreateGroupSchedulePage from "@pages/CreateGroupSchedulePage";
 import CreateMySchedulePage from "@pages/CreateMySchedulePage/page";
 import ErrorPage from "@pages/ErrorPage/page";
@@ -11,7 +11,6 @@ import FriendManagementPage from "@pages/FriendManagementPage/page";
 import GroupCalenderPage from "@pages/GroupCalenderPage/page";
 import GroupListPage from "@pages/GroupListPage/page";
 import GroupMemberPage from "@pages/GroupMemberPage/page";
-import GroupPage from "@pages/GroupPage/page";
 import GroupScheduleDetailPage from "@pages/GroupScheduleDetailPage/page";
 import InvitingPage from "@pages/InvitingPage/page";
 import LocationSharingPage from "@pages/LocationSharingPage/page";
